#199: extract group element serializer;
