+++ src/main/scala/sigmastate/serialization/DataSerializer.scala
@@ -3,20 +3,13 @@ package sigmastate.serialization
 import java.math.BigInteger
 
 import org.ergoplatform.ErgoBox
-import scorex.crypto.authds.ADDigest
-import sigmastate.SCollection.SByteArray
 import sigmastate.utils.{ByteWriter, ByteReader}
 import sigmastate._
-import sigmastate.interpreter.CryptoConstants
 import sigmastate.interpreter.CryptoConstants.EcPointType
 import scala.collection.mutable
 
 /** This works in tandem with ConstantSerializer, if you change one make sure to check the other.*/
 object DataSerializer {
-  type ElemType = CryptoConstants.EcPointType
-
-  private val curve = CryptoConstants.dlogGroup
-  private val LengthSize: Int = 2
 
   def serialize[T <: SType](v: T#WrappedType, tpe: T, w: ByteWriter): Unit = tpe match {
     case SUnit => // don't need to save anything
@@ -32,14 +25,11 @@ object DataSerializer {
       w.putUShort(length.toShort)
       w.putBytes(data)
     case SGroupElement =>
-      val bytes = v.asInstanceOf[EcPointType].getEncoded(true)
-      w.putBytes(bytes)
+      GroupElementSerializer.serializeBody(v.asInstanceOf[EcPointType], w)
     case SBox =>
       ErgoBox.serializer.serializeBody(v.asInstanceOf[ErgoBox], w)
-
     case SAvlTree =>
       AvlTreeData.serializer.serializeBody(v.asInstanceOf[AvlTreeData], w)
-
     case tCol: SCollectionType[a] =>
       val arr = v.asInstanceOf[tCol.WrappedType]
       val len = arr.length
@@ -78,20 +68,7 @@ object DataSerializer {
       val valueBytes = r.getBytes(size)
       new BigInteger(valueBytes)
     case SGroupElement =>
-      r.getByte() match {
-        case 0 =>
-          // infinity point is always compressed as 1 byte (X9.62 s 4.3.6)
-          val point = curve.curve.decodePoint(Array(0)).asInstanceOf[ElemType]
-          point
-        case m if m == 2 || m == 3 =>
-          val consumed = 1 + (curve.curve.getFieldSize + 7) / 8
-          r.position = r.position - 1
-          val encoded = r.getBytes(consumed)
-          val point = curve.curve.decodePoint(encoded).asInstanceOf[ElemType]
-          point
-        case m =>
-          throw new Error(s"Only compressed encoding is supported, $m given")
-      }
+      GroupElementSerializer.parseBody(r)
     case SBox =>
       ErgoBox.serializer.parseBody(r)
     case SAvlTree =>
